// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io

export interface LoginData {
    email: string;
    password: string;
}

export interface RegisterData {
    fullName: string;
    email: string;
    password: string;
}

export interface User {
    id:        string;
    email:     string;
    password:  string;
    token:     string;
    fullName?: string;
}


export interface RegisterResponse {
    email:    string;
    fullName: string;
    id:       string;
    token:    string;
}